backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
yt-dlp==2023.12.30
opencv-python==4.9.0.80
numpy==1.26.4
google-genai==0.3.0
